Revert "Update iconv commit to latest to support go1.6 cgo strict rule"
This reverts commit e4d361c5ab90c8b22e6b67604e7915c81cdab569.
The above commit is not working properly as it will convert wrongly
characters. Seem to be caused by improper use pointer because iconv
expecting pointer to pointer which it seem to allocate the memory
internally. This allocated memory is not being passed out to
golang side as the converted characters is not visible to golang side.
This is a quick fix before proper inspection should be done on iconv
pkg. However, without that commit mean that this pkg can not support
go1.6 and above as document here:
https://github.com/golang/go/issues/12416