Update 'cention-sany/gomail' Pkg To Latest To Fix *.eml Attachment
Any attachment with *.eml extension will likely send as
content-type message/rfc822 which can not use base64 transfer
encoding scheme according to RFC spec. But this library always use
base64. This commit make the sending of *.eml attachment always
send as attachment with content-type application instead of
message/rfc822 so that the limitation using base64 is not there.