CS-2052 Email out to Gmail shows BCC recipients
Added BCC field to message struct and removed retrieving of
BCC field from headers when building recipient list.
This is to prevent BCC from showing up in the DATA section
of smtp which will cause the other recipient to have the list
of bcc recipients.
the main branch of this fork has this problem fix but at the moment
the changes are deemed incompatible with this branch